refactor and move RMSNorm override
