+++ integrations/vllm_plugin/vllm_tt/vllm_distributed_utils.py
@@ -314,103 +314,3 @@ def shard_model(model: torch.nn.Module, mesh: "xs.Mesh") -> None:
             _process_module(child_module, child_name, module)
 
     _process_module(model)
-
-
-class TTRMSNorm(nn.Module):
-    """TT-compatible RMSNorm replacement for vLLM's RMSNorm.
-
-    vLLM's RMSNorm.forward_native accesses `self.weight.data`, which causes an
-    AssertionError during torch.compile/torch.export tracing with FakeTensors.
-    Accessing `.data` on a FakeTensor lifts it out of the fake tensor context,
-    resulting in: "cannot call `.data` on a Tensor, the Tensor is a FakeTensor".
-
-    This class reimplements the RMSNorm forward pass using `self.weight` directly
-    (without `.data`), making it compatible with TT tracing and compilation.
-    """
-
-    def __init__(self, layer: nn.Module):
-        super().__init__()
-        assert isinstance(layer, RMSNorm)
-        self.hidden_size = layer.hidden_size
-        self.variance_epsilon = layer.variance_epsilon
-        self.variance_size_override = layer.variance_size_override
-        self.has_weight = layer.has_weight
-        self.weight = layer.weight
-
-        if hasattr(layer, "rocm_norm_func") and hasattr(
-            layer, "rocm_norm_func_with_add"
-        ):
-            self.rocm_norm_func = layer.rocm_norm_func
-            self.rocm_norm_func_with_add = layer.rocm_norm_func_with_add
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        residual: torch.Tensor | None = None,
-    ) -> torch.Tensor | tuple[torch.Tensor, torch.Tensor]:
-
-        orig_dtype = x.dtype
-        x = x.to(torch.float32)
-        if residual is not None:
-            # residual promoted f16->f32 automatically,
-            # otherwise Inductor eliminates the casts to and from f16,
-            # increasing memory usage (and complicating pattern matching)
-            x = x + residual
-            residual = x.to(orig_dtype)
-
-        if x.shape[-1] != self.hidden_size:
-            raise ValueError(
-                f"Expected hidden_size to be {self.hidden_size}, but found: {x.shape[-1]}"
-            )
-
-        if self.variance_size_override is None:
-            x_var = x
-        else:
-            if self.hidden_size < self.variance_size_override:
-                raise ValueError(
-                    "Expected hidden_size to be at least "
-                    f"{self.variance_size_override}, but found: {self.hidden_size}"
-                )
-
-            x_var = x[:, :, : self.variance_size_override]
-
-        variance = x_var.pow(2).mean(dim=-1, keepdim=True)
-
-        x = x * torch.rsqrt(variance + self.variance_epsilon)
-        x = x.to(orig_dtype)
-        if self.has_weight and self.weight is not None:
-            x = x * self.weight
-        if residual is None:
-            return x
-        else:
-            return x, residual
-
-
-def tt_rmsnorm_module(layer: torch.nn.Module) -> torch.nn.Module:
-    assert isinstance(layer, RMSNorm)
-    tt_layer = TTRMSNorm(layer)
-    # logger.info("Wrapped RMSNorm module %s with XLA-compatible version", layer)
-    return tt_layer
-
-
-def replace_rmsnorm_modules(model: torch.nn.Module) -> None:
-    def _process_module(module, name=None, parent=None):
-        if get_fqn(module) == "RMSNorm":
-            wrapped_module = tt_rmsnorm_module(module)
-
-            assert (
-                parent is not None and name is not None
-            ), "Top Level module is not expected to be wrapped."
-            if wrapped_module is not module:
-                # Wrapped module and module are different py object.
-                # The original module should be replaced by the
-                # wrapped_module.
-                logger.debug("replace %s with %s", module, wrapped_module)
-                setattr(parent, name, wrapped_module)
-
-            module = wrapped_module
-
-        for child_name, child_module in list(module.named_children()):
-            _process_module(child_module, child_name, module)
-
-    _process_module(model)
